mpnn/mpnn_jax/requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
chex==0.1.91
contourpy==1.3.3
cycler==0.12.1
etils==1.13.0
flax==0.12.0
fonttools==4.60.1
fsspec==2025.10.0
humanize==4.14.0
importlib_resources==6.5.2
jax==0.8.0
jaxlib==0.8.0
jraph==0.0.6.dev0
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
msgpack==1.1.2
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
opt_einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.28
packaging==25.0
pillow==12.0.0
protobuf==6.33.0
psutil==7.1.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rich==14.2.0
scipy==1.16.3
simplejson==3.20.2
six==1.17.0
tensorstore==0.1.79
toolz==1.1.0
tqdm==4.67.1
treescope==0.1.10
typing_extensions==4.15.0
zipp==3.23.0
